Update VimLab.
add quick fix list feature to look at the errors. Basically it create a temporary .m file called MatVimTmp.m in the same folder and run it and then catch the errors which are saved in the /tmp/MatlabErrorForVim.err. when press \mq, the errors are loaded by calling cfile('/tmp/MatlabErrorForVim.err')